ISSD-1654 Move parent folder and format date
