+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedAppsDashboard/Apps/App/AppOutlet.tsx
@@ -101,7 +101,7 @@ const AppOutlet = () => {
 
 			<AppNavbar />
 
-			<Outlet context={{data}} />
+			<Outlet context={data} />
 		</div>
 	);
 };
